2 controllers

+++ antifraud/src/main/java/com/bank/antifraud/repository/SuspiciousAccountTransferRepository.java
@@ -4,6 +4,9 @@ import com.bank.antifraud.model.SuspiciousAccountTransfer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+import java.util.List;
+
 @Repository
-public interface SuspiciousAccountTransferRepository extends JpaRepository<SuspiciousAccountTransfer, Integer> {
+public interface SuspiciousAccountTransferRepository extends JpaRepository<SuspiciousAccountTransfer, BigInteger> {
 }
